refactor: complete purifying Response module
